Allow token based gem fetching
By creating an encrypted and expiring message we can allow for sharing the
message over insecure transports IE. Docker images.

The idea is that the user will first request a "token" which is an encrypted
message containing the device id and a timestamp when the token was generated.
This token is unreadable to anyone but the instance that generated the token.

The user may then substitute regular credentials with the generated token for
use with Bundler to fetch gems from the gemserver.